🚀 Separate unsafe calls 🚀
Put method calls 🚀 in separate 🚀 unsafe blocks. This is done so that they don't come together and use their evil to blast out 🚀 of the unsafe block 🚀 and cause havoc. Of course, this is rather unlikely to happen anyway 🚀 because of rust's 🚀 amazing 🚀 safety 🚀, but it's always good 🚀 to be sure 🚀.